chore : 크롭 이미지 추출과 이미지 업로드 분리
